Added better relay support for controller fan switching
replaced CONTROLLERFAN_SPEED with CONTROLLERFAN_OFF and
CONTROLLERFAN_ON. It still uses the same 0-255 value range, and will
still behave the same way when simply selecting a pin for switching,

Since the 5v signal will be used primarily for switching, this allows
for setting the on state as needed for relays that require Active Low
signals for activation.

It replaces the hardcoded 0 (off) value with CONTROLLERFAN_OFF which can
be defined by the user.